+++ docs/nuxt.config.js
@@ -2,6 +2,6 @@ import theme from '@nuxt/content-theme-docs'
 
 export default theme({
   docs: {
-    primaryColor: '#ae9dff'
-  }
+    primaryColor: '#ae9dff',
+  },
 })