+++ test/fixture/default/pages/index.vue
@@ -4,6 +4,11 @@
     <span id="server-side">{{ serverSentry ? 'Works!' : '$sentry object is missing!' }}</span>
     <h3>Client-side</h3>
     <span id="client-side">{{ clientSentry ? 'Works!' : '$sentry object is missing!' }}</span>
+    <p>
+      <button id="crash-button" @click="crash_me()">
+        crash me
+      </button>
+    </p>
   </div>
 </template>
 
@@ -11,9 +16,9 @@
 export default {
   data () {
     return {
-      /** @type {import('@sentry/minimal') | null} */
+      /** @type {import('@sentry/core') | null} */
       clientSentry: null,
-      /** @type {import('@sentry/minimal') | null} */
+      /** @type {import('@sentry/core') | null} */
       serverSentry: this.$sentry,
     }
   },